fix dtype bug

+++ tfdet/core/anchor/generator.py
@@ -146,7 +146,7 @@ def generate_yolo_anchors(feature, image_shape = [608, 608], size = [[0.01645, 0
         out = tf.concat(out, axis = 0)
     return out
 
-def generate_points(feature, image_shape = [1024, 1024], stride = None, normalize = True, flatten = True, concat = True):
+def generate_points(feature, image_shape = [1024, 1024], stride = None, normalize = True, flatten = True, concat = True, dtype = tf.float32):
     if tf.is_tensor(feature) or not isinstance(feature, list) or isinstance(feature[0], int):
         feature = [feature]
     if not isinstance(stride, list):
